+++ frontend/rust-lib/flowy-document2/src/event_handler.rs
@@ -7,8 +7,7 @@
 use std::sync::{Arc, Weak};
 
 use collab_document::blocks::{
-  json_str_to_hashmap, Block, BlockAction, BlockActionPayload, BlockActionType, BlockEvent,
-  BlockEventPayload, DeltaType,
+  BlockAction, BlockActionPayload, BlockActionType, BlockEvent, BlockEventPayload, DeltaType,
 };
 
 use flowy_error::{FlowyError, FlowyResult};
@@ -250,24 +249,6 @@ impl From<BlockActionPayloadPB> for BlockActionPayload {
       prev_id: pb.prev_id,
       text_id: pb.text_id,
       delta: pb.delta,
-    }
-  }
-}
-
-impl From<BlockPB> for Block {
-  fn from(pb: BlockPB) -> Self {
-    // Use `json_str_to_hashmap()` from the `collab_document` crate to convert the JSON data to a hashmap
-    let data = json_str_to_hashmap(&pb.data).unwrap_or_default();
-
-    // Convert the protobuf `BlockPB` to our internal `Block` struct
-    Self {
-      id: pb.id,
-      ty: pb.ty,
-      children: pb.children_id,
-      parent: pb.parent_id,
-      data,
-      external_id: pb.external_id,
-      external_type: pb.external_type,
     }
   }
 }
